% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreGated.R
\name{zscoreGated}
\alias{zscoreGated}
\title{5 level Gated Zscore Function}
\usage{
zscoreGated(df, colname)
}
\arguments{
\item{df}{a dataframe for scoring}

\item{colname}{variable name for scoring}
}
\value{
a dataframe with gated score attached
}
\description{
This function allows you to calculated z-scores and 5 level gated scores
}
\examples{
zscoreGated(dataframe,"math")
}
